requirements.txt
numpy
opencv-python
pillow
tensorflow        # if using TF model (optional)
streamlit
flask
werkzeug